split monolithic service into seperate services
