+++ order/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "order",
+  "version": "1.0.0",
+  "description": "The Orders service handles the purchasing process.",
+  "license": "ISC",
+  "author": "",
+  "type": "commonjs",
+  "main": "server.js",
+  "scripts": {
+    "start": "nodemon dist/server.js",
+    "dev": "nodemon src/server.ts",
+    "test": "jest"
+  },
+  "devDependencies": {
+    "@types/express": "^5.0.0",
+    "@types/node": "^22.13.1",
+    "nodemon": "^3.1.9",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.7.3"
+  },
+  "dependencies": {
+    "cors": "^2.8.5",
+    "dotenv": "^16.4.7",
+    "express": "^4.21.2",
+    "uuid": "^11.0.5"
+  }
+}
