+++ user/src/routes/userRoutes.ts
@@ -1,9 +1,9 @@
 import express, { Request, Response } from "express";
-import { createOne, findOneByEmail } from "../mongoDatabase/user";
+// import { createOne, findOneByEmail } from "../mongoDatabase/user";
 import { v4 as uuidv4 } from "uuid";
 import * as User from "../mongoDatabase/user";
-import * as Order from "../mongoDatabase/order";
-import * as Product from "../mongoDatabase/product";
+// import * as Order from "../../order/src/mongoDatabase/order";
+// import * as Product from "../mongoDatabase/product";
 const router = express.Router();
 
 router.post("/register", async (req: Request, res: Response): Promise<any> => {
@@ -17,7 +17,7 @@ router.post("/register", async (req: Request, res: Response): Promise<any> => {
     return res.status(400).json({ error: "Invalid email format" });
   }
 
-  const existingUser = findOneByEmail(email);
+  const existingUser = User.findOneByEmail(email);
   if (existingUser) {
     return res.status(400).json({ error: "User already exists" });
   }
@@ -33,7 +33,7 @@ router.post("/register", async (req: Request, res: Response): Promise<any> => {
     isActive: true,
     role: "user" as const,
   };
-  createOne(newUser);
+  User.createOne(newUser);
 
   console.log(`Sending welcome email to ${email}`);
 
@@ -42,26 +42,26 @@ router.post("/register", async (req: Request, res: Response): Promise<any> => {
     .json({ message: "User registered successfully", user: newUser });
 });
 
-router.get("/:id/orders", async (req: Request, res: Response): Promise<any> => {
-  const { id } = req.params;
+// router.get("/:id/orders", async (req: Request, res: Response): Promise<any> => {
+//   const { id } = req.params;
 
-  const user = User.findOneById(id);
-  if (!user) {
-    return res.status(404).json({ error: "User not found" });
-  }
+//   const user = User.findOneById(id);
+//   if (!user) {
+//     return res.status(404).json({ error: "User not found" });
+//   }
 
-  const userOrders = Order.findAllByUserId(id);
+//   const userOrders = Order.findAllByUserId(id);
 
-  const detailedOrders = userOrders.map((order) => {
-    const product = Product.findOneById(order.productId);
-    return {
-      ...order,
-      productName: product ? product.name : "Unknown",
-      productPrice: product ? product.price : 0,
-    };
-  });
+//   const detailedOrders = userOrders.map((order) => {
+//     const product = Product.findOneById(order.productId);
+//     return {
+//       ...order,
+//       productName: product ? product.name : "Unknown",
+//       productPrice: product ? product.price : 0,
+//     };
+//   });
 
-  res.json(detailedOrders);
-});
+//   res.json(detailedOrders);
+// });
 
 export default router;
