+++ user/src/server.ts
@@ -1,8 +1,8 @@
 import express, { Request, Response, Router } from "express";
 import dotenv from "dotenv";
 import userRoutes from "./routes/userRoutes";
-import productRoutes from "./routes/productRoutes";
-import orderRoutes from "./routes/orderRoutes";
+// import productRoutes from "./routes/productRoutes";
+// import orderRoutes from "./routes/orderRoutes";
 
 dotenv.config();
 const app = express();
@@ -10,12 +10,12 @@ const app = express();
 app.use(express.json());
 
 app.use("/users", userRoutes);
-app.use("/products", productRoutes);
-app.use("/orders", orderRoutes);
+// app.use("/products", productRoutes);
+// app.use("/orders", orderRoutes);
 
 app.get("/", (req: Request, res: Response) => {
   res.status(200).json({ message: "Hello World" });
 });
 
-const PORT = process.env.PORT || 3000;
+const PORT = process.env.PORT || 3002;
 app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
